+++ packages/cli/src/download/data-set/DataSet.handler.ts
@@ -36,6 +36,7 @@ export default class DownloadDataSetHandler extends SshBaseHandler {
             stream: fs.createWriteStream(localFilePath),
             dsname: params.arguments.dataSet,
             encoding: params.arguments.binary ? "binary" : params.arguments.encoding,
+            localEncoding: params.arguments.localEncoding,
         });
 
         params.response.data.setMessage("Successfully downloaded content to %s", localFilePath);